Add backend methods to DiscussionService, DiscussionPost and post-route for updating a post's image. Add incomplete implementation of 'edit' link to React PostLinks component

+++ pictochat-be/routes/post-route.ts
@@ -2,10 +2,18 @@ import util from 'util';
 import fs from 'fs';
 import express from 'express';
 import multer from 'multer';
-import { DiscussionService } from '../services/discussion-service';
+import passport from 'passport';
+import { strategies } from '../middleware/passport-middleware';
+import { DiscussionService, NewReply, NewThread } from '../services/discussion-service';
 import { MIMETYPE_TO_ENCODING } from '../utils/encoding-content-types';
 import { DiscussionTreeNode } from '../models/discussion-tree-node';
 import config from '../utils/config';
+import { User } from '../models/user';
+import {ForbiddenError} from '../exceptions/forbidden-error';
+import * as ErrorUtils from '../exceptions/error-utils';
+import { DiscussionPost } from '../models/discussion-post';
+
+const AUTHENTICATION_ERROR = 'AUTHENTICATION_ERROR';
 
 //// HELPER FUNCTIONS ////
 
@@ -19,27 +27,47 @@ async function makeNewImageSpec(file): Promise<{ data: Buffer; encoding: string 
   return { data, encoding };
 }
 
+function assertIsPostAuthor(body: NewReply | NewThread, user: User) {
+  if (body.userId !== user.userId) {
+    throw new ForbiddenError("Post's userId and/or supplied JWT token is incorrect or are for different users");
+  }
+}
+
 async function handleNewReplyPOST(req, res, next) {
-  let newImageSpec = await makeNewImageSpec(req.file);
-  let newReplySpec = {
-    image: newImageSpec,
-    parentPostId: req.body.parentPostId,
-    userId: req.body.userId
-  };
-  let post = await DiscussionService.createReply(newReplySpec);
-  res.json(post.toJSON());
-  await deleteFile(req.file.path);
+  try {
+    assertIsPostAuthor(req.body, req.user);
+    let newImageSpec = await makeNewImageSpec(req.file);
+    let newReplySpec = {
+      image: newImageSpec,
+      parentPostId: req.body.parentPostId,
+      userId: req.body.userId
+    };
+    let post = await DiscussionService.createReply(newReplySpec);
+    // Setting Location and using status 201 to match RESTful conventions for POST responses
+    res.set('Location', `${config.API_ROOT}/post/${post.postId}`);
+    res.status(201);
+    res.json(post.toJSON());
+  } finally {
+    await deleteFile(req.file.path);
+  }
 }
 
 async function handleNewThreadPOST(req, res, next) {
+  try {
+  assertIsPostAuthor(req.body, req.user);
   let newImageSpec = await makeNewImageSpec(req.file);
   let newThreadSpec = { image: newImageSpec, userId: req.body.userId };
   let thread = await DiscussionService.createThread(newThreadSpec);
+  // Setting Location and using status 201 to match RESTful conventions for POST responses
+  res.set('Location', `${config.API_ROOT}/post/${thread.rootPost.postId}`);
+  res.status(201);
   res.json(thread.toFlatJSON());
-  await deleteFile(req.file.path);
+  } finally {
+    await deleteFile(req.file.path);
+  }
 }
 
-// This will store images in a local staging directory on the API server
+// This will temporarily store images in a local staging directory on the API server
 const imageStager = multer({ dest: config.IMAGE_STAGING_DIR });
 
 //// ROUTER ////
@@ -58,14 +86,45 @@ postRouter.get('/:postId', async (req, res, next) => {
   }
 });
 
-postRouter.post('/', imageStager.single('image'), async (req, res, next) => {
-  try {
-    if (!!req.body.parentPostId) {
-      await handleNewReplyPOST(req, res, next);
-    } else {
-      await handleNewThreadPOST(req, res, next);
+postRouter.post('/',
+  passport.authenticate(strategies.JWT, { session: false }),
+  imageStager.single('image'),
+  async (req, res, next) => {
+    try {
+      if (!!req.body.parentPostId) {
+        await handleNewReplyPOST(req, res, next);
+      } else {
+        await handleNewThreadPOST(req, res, next);
+      }
+    } catch (error) {
+      next(error);
     }
-  } catch (error) {
-    next(error);
   }
-});
+);
+
+// Update a post
+postRouter.patch('/:postId',
+  passport.authenticate(strategies.JWT, { session: false }),
+  imageStager.single('image'),
+  async (req: any, res, next) => {
+    try {
+      let newImageSpec = await makeNewImageSpec(req.file);
+      let postUpdateSpec = {
+        image: newImageSpec,
+        postId: req.body.postId,
+        userId: req.user.userId
+      };
+      let post: DiscussionPost = await DiscussionService.updatePost(postUpdateSpec);
+      // Return full tree under updated post so that clients only have to deal with one
+      // data structure for all methods.
+      let postTree = await DiscussionService.getReplyTreeUnderPost(post.postId);
+      res.status(200);
+      res.json(postTree.toJSON());
+    } catch (error) {
+      next(error);
+    } finally {
+      deleteFile(req.file.path);
+    }
+  }
+);
+
